refactor: Move MultiselectBackdrop in own component
No code change.
